Fix : onboarding, username 중복 확인 API 수정

+++ src/main/java/capston/new_valance/repository/TagRepository.java
@@ -5,6 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TagRepository extends JpaRepository<Tag, Long> {
-    Optional<Tag> findByTagName(String tagName);
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
+
+public interface TagRepository extends JpaRepository<Tag, Integer> {
+    Optional<Tag> findByTagName(String tagName); // 태그 이름으로 태그 조회
 }
